+++ app/controllers/streams_controller.rb
@@ -64,8 +64,8 @@ class StreamsController < ApplicationController
     end
 
     respond_with do |format|
-      format.html { render 'layouts/main_stream' }
-      format.mobile { render 'layouts/main_stream' }
+      format.html { render 'streams/main_stream' }
+      format.mobile { render 'streams/main_stream' }
       format.json { render :json => @stream.stream_posts.map {|p| LastThreeCommentsDecorator.new(PostPresenter.new(p, current_user)) }}
     end
   end