+++ Changelog.md
@@ -11,6 +11,7 @@
 ## Refactor
 
 * Refactor people_controller#show and photos_controller#index [#4002](https://github.com/diaspora/diaspora/issues/4002)
+* Modularize layout [#3944](https://github.com/diaspora/diaspora/pull/3944)
 
 ## Features